upgrade all components to TS
