+++ src/routes/albums/album-details.js
@@ -1,7 +1,8 @@
 import { html, LitElement } from 'lit';
 import { unsafeHTML } from 'lit/directives/unsafe-html.js';
 import { getAlbumById } from '../../services/albums-service.js';
-import { modelAlbum } from '../../components/card/card.js';
+import { modelAlbum } from '../../components/card/card.model.ts';
+import '../../components/card/card.ts';
 import albumsCss from './albums.css?type=css';
 
 class AlbumDetailsRouteComponent extends LitElement {
